# api/openapi.yaml
openapi: 3.0.3
...
info:
  title: LinkCtrl API
  version: "1.0"
  description: |
    Self-hostable link management. Every capability of the dashboard is
    available here; the two surfaces call the same service layer.

    ### Authentication
    Two credentials work everywhere:
    * **API key** — `Authorization: Bearer lk_live_…`. Keys are minted in the
      dashboard or with `lctl apikey create`, and act as their owner limited to
      the scopes granted at creation, intersected with the owner's current
      role on every request.
    * **Session cookie** — set by `POST /auth/login`, primarily for the
      dashboard. Cookie-authenticated unsafe requests must be same-origin.

    If a request carries both, the bearer key wins.

    ### Errors
    Every error is an RFC 9457 problem document
    (`application/problem+json`). Validation failures carry a per-field
    `errors` array shaped for form rendering.

    ### Rate limits
    **Any** operation here may answer `429` with a `Retry-After` header once
    the per-address API limit is exceeded, so it is documented once rather
    than repeated on every path; the endpoints that verify a credential carry
    a second, much tighter limit and list it explicitly. Two problem types
    return `429` and mean different things: `rate-limited` is "you are going
    too fast", `account-locked` is "this account is frozen for a while".
    Retrying the first is fine; retrying the second is not.

...
    per-request deadline.

    ### Privacy
...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
  /auth/setup:
    post:
      tags: [auth]
      operationId: setupInstance
      summary: Claim a fresh instance by creating the first account
      description: |
        Exists only while the instance has no users, then returns 404 forever.
        The first account is an owner and skips email verification: whoever can
        reach a fresh instance's setup endpoint deployed it.
      security: []
      requestBody:
        $ref: "#/components/requestBodies/Credentials"
      responses:
        "201":
          description: Instance claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountCreated"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Validation"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Create an account
      description: Subject to the instance's signup mode; new instances are closed.
      security: []
      requestBody:
        $ref: "#/components/requestBodies/Credentials"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountCreated"
        "403":
          description: Registration is closed on this instance
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Validation"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Sign in and receive a session cookie
      description: |
        Every credential failure returns the same 401, whatever the cause —
        distinguishing them would reveal which addresses are registered.

        Two independent limits guard this endpoint, and both answer 429 with
        different problem types: repeated failures against one account lock it
        briefly (`account-locked`), and too many attempts from one address are
        throttled whatever accounts they name (`rate-limited`). The second is
        what credential stuffing across a leaked list looks like, and the first
        would never see it.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                name:
                  type: string
                  description: Ignored on login; accepted so one credentials shape serves register and login.
      responses:
        "200":
          description: Signed in; the session cookie is on the response
          headers:
            Set-Cookie:
              schema:
                type: string
              description: "`__Host-linkctrl_session` (unprefixed over plain HTTP in development)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/Validation"
        "429":
          description: |
            Either the account is temporarily locked after repeated failures
            (`account-locked`), or this address is being throttled
            (`rate-limited`). `Retry-After` is present in the second case.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /auth/password:
    post:
      tags: [auth]
      operationId: changePassword
      summary: Change the password and sign out every other session
      description: |
        Requires an interactive session — an API key is refused, because a
        leaked key must not be able to lock its owner out. The current session
        survives; all others are revoked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 12
      responses:
        "204":
          description: Password changed, other sessions revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"
        "429":
          $ref: "#/components/responses/RateLimited"
  /me:
    get:
      tags: [auth]
      operationId: getMe
      summary: The authenticated identity and its effective permissions
      description: |
        For an API key, `permissions` is the key's scopes intersected with the
        owner's current role — what this request can actually do, not what was
        granted at creation.
      responses:
        "200":
          description: The current identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /links:
    get:
      tags: [links]
      operationId: listLinks
      summary: List links with search, filters and keyset pagination
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Full-text over title/description, plus substring on alias and URL.
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LinkStatus"
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, clicks]
            default: newest
        - name: cursor
          in: query
          schema: { type: string }
          description: Opaque, from `next_cursor` of the previous page.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1 }
          description: Page size; out-of-range values fall back to the server default.
        - name: include_total
          in: query
          schema: { type: boolean }
          description: Counting costs a scan, so the total is opt-in.
        - name: tag
          in: query
          schema:
            type: array
            items: { type: string, format: uuid }
          style: form
          explode: true
          description: Repeatable; links matching any given tag id.
      responses:
        "200":
          description: One page of links
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [links]
      operationId: createLink
      summary: Create a link
      description: |
        Destinations are restricted to http/https and may not point at
        private, loopback, link-local or cloud-metadata addresses. The
        `password`, `max_clicks` and `one_time` fields exist in the schema but
        are rejected with a 422 until Phase 2 — refusing loudly beats
        accepting and doing nothing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201":
          description: Link created
          headers:
            Location:
              schema: { type: string }
              description: The public short URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Validation"
  /links/{id}:
    parameters:
      - $ref: "#/components/parameters/LinkID"
    get:
      tags: [links]
      operationId: getLink
      summary: Fetch one link
      responses:
        "200":
          description: The link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [links]
      operationId: updateLink
      summary: Edit a link without changing its identity
      description: |
        Partial update: absent fields are untouched. `expires_at` is
        three-valued — absent leaves the expiry alone, an empty string clears
        it, a timestamp sets it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkRequest"
      responses:
        "200":
          description: The updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Validation"
    delete:
      tags: [links]
      operationId: deleteLink
      summary: Soft-delete a link
      description: |
        The link stops redirecting immediately and stays restorable for 30
        days before it is purged. An alias that received traffic is never
        reissued after the purge.
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /links/{id}/archive:
    post:
      tags: [links]
      operationId: archiveLink
      summary: Archive a link
      description: Stops redirecting but keeps the alias reserved and the analytics visible.
      parameters:
        - $ref: "#/components/parameters/LinkID"
      responses:
        "200":
          description: The archived link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /links/{id}/restore:
    post:
      tags: [links]
      operationId: restoreLink
      summary: Restore an archived link
      parameters:
        - $ref: "#/components/parameters/LinkID"
      responses:
        "200":
          description: The restored link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /links/{id}/stats:
    get:
      tags: [analytics]
      operationId: getLinkStats
      summary: A link's analytics over a date window
      parameters:
        - $ref: "#/components/parameters/LinkID"
        - $ref: "#/components/parameters/FromDay"
        - $ref: "#/components/parameters/ToDay"
      responses:
        "200":
          description: Totals, a day series and per-dimension breakdowns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Validation"
  /links/{id}/clicks:
    get:
      tags: [analytics]
      operationId: getLinkClicks
      summary: Recent clicks on a link
      description: |
        The one endpoint that reads raw events rather than rollups; bounded
        and index-backed, so its cost does not grow with history. No addresses
        appear because none are stored.
      parameters:
        - $ref: "#/components/parameters/LinkID"
      responses:
        "200":
          description: The most recent clicks, newest first
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecentClick"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /stats/overview:
    get:
      tags: [analytics]
      operationId: getOverview
      summary: Workspace-wide totals and day series
      parameters:
        - $ref: "#/components/parameters/FromDay"
        - $ref: "#/components/parameters/ToDay"
      responses:
        "200":
          description: The workspace overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Overview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/Validation"
  /tags:
    get:
      tags: [tags]
      operationId: listTags
      summary: List the workspace's tags with usage counts
      responses:
        "200":
          description: All tags
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tag"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tags/{id}:
    delete:
      tags: [tags]
      operationId: deleteTag
      summary: Delete a tag, detaching it from every link
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api-keys:
    get:
      tags: [api-keys]
      operationId: listAPIKeys
      summary: List your own API keys
      description: |
        Requires an interactive session. Revoked keys stay listed until a
        reaper removes them, because "which keys existed" is the question
        asked after an incident. The secret is never listed — only its hash
        is stored.
      responses:
        "200":
          description: Your keys, newest first
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [api-keys]
      operationId: createAPIKey
      summary: Mint an API key
      description: |
        Requires an interactive session — a key cannot mint keys, or revoking
        a leaked one would be meaningless. Scopes must be permissions you
        yourself hold; key management and organization deletion are never
        grantable. The response is the only time the token is shown.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: The key, including the token that will never be shown again
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedAPIKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"
  /api-keys/{id}:
    delete:
      tags: [api-keys]
      operationId: revokeAPIKey
      summary: Revoke an API key
      description: |
        Takes effect on the key's next request; nothing about keys is cached.
        Someone else's key returns 404, not 403, so ids cannot be probed.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Revoked (idempotent)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API key, `lk_live_…`. Scopes are intersected with the owner's role per request.
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-linkctrl_session
      description: Set by `POST /auth/login`. Unsafe cross-origin requests are refused.
...
  requestBodies:
    Credentials:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [email, password]
            properties:
              email:
                type: string
                format: email
              password:
                type: string
                minLength: 12
                description: Length is the only rule (NIST SP 800-63B).
              name:
                type: string
                description: Display name; defaults to the mailbox part of the email.
  responses:
...
    NotFound:
      description: Does not exist, or belongs to a workspace this identity cannot see
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The alias or resource already exists
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Validation:
      description: One or more fields are invalid; `errors` names each one
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: |
        Too many requests from this address. `Retry-After` carries the number of
        seconds to wait; it is a floor, not a promise.
      headers:
        Retry-After:
          description: Seconds until a request from this address will be accepted
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details. Clients should branch on `type`, not prose.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          description: Present on validation failures; one entry per invalid field.
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
    AccountCreated:
      type: object
      additionalProperties: false
      required: [user_id, email, workspace_id]
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        workspace_id:
          type: string
          format: uuid
          description: The auto-provisioned personal workspace.
    SessionCreated:
      type: object
      additionalProperties: false
      required: [user_id, email, workspace_id, role, expires_at]
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        workspace_id: { type: string, format: uuid }
        role: { type: string }
        expires_at:
          type: string
          format: date-time
          description: The session's absolute deadline; idle expiry may end it sooner.
...